+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 
 - **feat**: enhance path handling for moved workspace in problem data
 - **feat**: add javascript language support
+- **fix**: show panel not working in new opening mode
 
 ## 0.4.8